requirements.txt
Babel==1.3
Django==1.8.2
Jinja2==2.7.3
MarkupSafe==0.23
Pygments==2.0.2
alabaster==0.7.4
argparse==1.2.1
dj-database-url==0.3.0
django-extensions==1.5.5
docutils==0.12
psycopg2==2.6
pytz==2015.4
six==1.9.0
snowballstemmer==1.2.0
wsgiref==0.1.2
pycrypto==2.6.1
